// src/app/data/model/profeDataGetTema.ts
export interface ProfeDataGetTema {
    cdicurso: string;
    id_curso: string;
    nombre_curso: string;
    cdiprofe: string;
    nombre_profe: string;
    cdicategoria: string;
    nombre_categoria: string;
    cditema: string;
    nombre_tema: string;
    descripcion_tema: string;
    cditest: string;
    traducir: string;
    idioma: string;
    ayuda: string;
    autocorreccion: string;
}